\name{caution}
\alias{caution}
\docType{data}
\title{Caution data}
\description{
Artificial data to illustrate problems with residual plots.  
}
\format{This data frame contains the following columns:
 \describe{
    \item{x1}{
Artificial data item.
    }
    \item{x2}{
Artificial data item.
    }
    \item{y}{
Artificial data item.
    }
 }
}

\source{
R. D. Cook and S. Weisberg (1999), Graphs in statistical analysis:  Is
the medium the message? \emph{American Statistician,}  53, 29-37.}
\references{Weirsbeg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Section 8.1.5.}
\examples{
head(caution)
}
\keyword{datasets}
