\name{florida}
\alias{florida}
\docType{data}
\title{Florida presidential election}
\description{
County-by-county vote for president in Florida in 2000 for Bush, Gore and Buchanan.
}
\format{
  A data frame three vaiaables for each of Florida's 67 counties.
  \describe{
    \item{Gore}{Vote for Gore}
    \item{Bush}{Vote for Bush}
    \item{Buchanan}{Vote for Buchanan}
  }
}
\source{
http://www.abcnews.go.com.sections/politics/2000vote/general/FL\_county.html
}
\references{
Weisberg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Problem 9.10.
}
\examples{
head(florida)
## maybe str(florida) ; plot(florida) ...
}
\keyword{datasets}
