\name{hooker}
\alias{hooker}
\docType{data}
\title{Hooker's data}
\description{
In his original paper, Forbes provided additional data collected by the botanist Dr. Joseph Hooker on temperatures and boiling points measured often at higher altitudes in the Himalaya Mountains.
}
\format{This data frame contains the following columns:
 \describe{
    \item{Temp}{
Measured boiling temperature, degrees F. 
    }
    \item{Pressure}{
Measured air pressure, inches of Mercury. 
    }
 }
}

\source{Forbes, J. (1957). Further experiments and remarks on the measurement of heights by boiling point of water. \emph{Transactions of the Royal Society of Edinburgh}, 21, 235-243.}
\references{Weisberg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Problem 2.2.4.}
\examples{
head(hooker)
}
\keyword{datasets}
