\name{stopping}
\alias{stopping}
\docType{data}
\title{Stopping distances}
\description{
Ezekiel and Fox (1959) data on auto stopping distances.
}
\format{This data frame contains the following columns:
 \describe{
    \item{Speed}{
 Speed (mph)
    }
    \item{Distance}{
 Stopping distance (in feet)
    }
 }
}

\source{Ezekiel, M. and Fox, K. A. (1959). \emph{Methods of Correlation Analysis, Linear and Curvilinear}, New York: Wiley.}
\references{Weisberg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Problem 6.2.}
\examples{
head(stopping)
}
\keyword{datasets}
