\name{turk0}
\alias{turk0}
\docType{data}
\title{Turkey data, one source}
\description{
  Turkey weight increase in an experiment in which the supplementation with methionine was
varied.
}
\format{This data frame contains the following columns:
 \describe{
    \item{A}{
 Amount of methionine supplement (percent of diet)
    }
    \item{Gain}{
 Pen weight increase (g)
    }
 }
}

\source{Cook, R. D. and Witmer, J. (1985). A note on the parameter-effects curvature. \emph{Journal of the American Statistical Association}, 80, 872-878.}
\references{Weisberg, S. (2005). \emph{Applied Linear Regression}, 3rd edition. New York: Wiley, Problem 6.13.}
\examples{
head(turk0)
}
\keyword{datasets}
